@@ -26,72 +26,72 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.4",
-    "@angular/cdk": "^7.3.2",
-    "@angular/common": "7.2.4",
-    "@angular/compiler": "7.2.4",
-    "@angular/core": "7.2.4",
-    "@angular/forms": "7.2.4",
-    "@angular/material": "^7.3.2",
-    "@angular/platform-browser": "7.2.4",
-    "@angular/platform-browser-dynamic": "7.2.4",
-    "@angular/router": "7.2.4",
-    "acorn": "^6.1.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "acorn": "^7.1.0",
     "body-parser": "^1.18.2",
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^4.4.0",
     "cookie-parser": "^1.4.3",
-    "core-js": "^2.4.1",
+    "core-js": "^3.4.4",
     "cors": "^2.8.4",
     "express": "^4.16.3",
-    "express-session": "^1.15.6",
-    "helmet": "^3.16.0",
+    "express-session": "^1.17.0",
+    "helmet": "^3.21.2",
     "hmacsha1": "^1.0.0",
-    "http-status-codes": "^1.3.0",
+    "http-status-codes": "^1.4.0",
     "intl": "^1.2.5",
     "jquery": "^3.3.1",
-    "json-pollock": "^1.2.0",
+    "json-pollock": "^1.3.0",
     "material-design-icons": "^3.0.1",
     "memorystore": "^1.6.1",
     "nconf": "^0.10.0",
     "ng-event-source": "^1.0.14",
-    "node": "^8.16.0",
+    "node": "^13.1.0",
     "node-rest-client": "^3.1.0",
-    "npm": "^6.9.0",
+    "npm": "^6.13.1",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.3",
     "sse-nodejs": "^0.1.0",
-    "url-polyfill": "^1.1.0",
-    "uuid": "^3.3.2",
+    "url-polyfill": "^1.1.7",
+    "uuid": "^3.3.3",
     "winston": "^3.0.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "7.3.1",
-    "@angular/compiler-cli": "7.2.4",
-    "@angular/language-service": "7.2.4",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "~0.803.19",
+    "@angular/cli": "8.3.19",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~12.12.14",
     "chai": "^4.2.0",
     "chai-http": "^4.2.1",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "karma-phantomjs-launcher": "^1.0.4",
-    "mocha": "^5.2.0",
-    "nock": "^10.0.3",
-    "nyc": "^13.3.0",
+    "mocha": "^6.2.2",
+    "nock": "^11.7.0",
+    "nyc": "^14.1.1",
     "phantomjs-prebuilt": "^2.1.16",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "3.2.4"
+    "ts-node": "~8.5.2",
+    "tslint": "~5.20.1",
+    "typescript": "~3.5.3"
   }
 }
