@@ -3,7 +3,10 @@ import {Component, EventEmitter, OnDestroy, OnInit, Output, ViewChild} from '@an
 import {Event} from '../../shared/models/app-installation/event.model';
 import {EndpointHeader} from "../../shared/models/app-installation/endpointHeaders.model";
 import {GeneralDetails} from "./lp-app-installation-general-details/GeneralDetails";
-import {MatDialog, MatSnackBar, MatStepper, MatTabGroup} from "@angular/material";
+import { MatDialog } from "@angular/material/dialog";
+import { MatSnackBar } from "@angular/material/snack-bar";
+import { MatStepper } from "@angular/material/stepper";
+import { MatTabGroup } from "@angular/material/tabs";
 import {AppInstallationsService} from "../../core/services/app-installations.service";
 import {AppInstall} from "../../shared/models/app-installation/appInstall.model";
 import {Webhooks} from "../../shared/models/app-installation/webhooks.model";
@@ -29,11 +32,11 @@ export class LpAppInstallationsComponent implements OnInit, OnDestroy {
   public avaliableEventTypes: object;
   public completed = false;
   @Output() eventCreated = new EventEmitter<Event>();
-  @ViewChild('tabs') tabs: MatTabGroup;
-  @ViewChild('stepperCreate') stepperCreate: MatStepper;
-  @ViewChild('stepperUpdate') stepperUpdate: MatStepper;
-  @ViewChild('appInstallGeneralDetails') appInstallGeneralDetails;
-  @ViewChild('updateAppInstallGeneralDeateils') updateAppInstallGeneralDeateils;
+  @ViewChild('tabs', {static: false}) tabs: MatTabGroup;
+  @ViewChild('stepperCreate', {static: false}) stepperCreate: MatStepper;
+  @ViewChild('stepperUpdate', {static: false}) stepperUpdate: MatStepper;
+  @ViewChild('appInstallGeneralDetails', {static: false}) appInstallGeneralDetails;
+  @ViewChild('updateAppInstallGeneralDeateils',  {static: false}) updateAppInstallGeneralDeateils;
   
   constructor(public appInstallationService: AppInstallationsService,
               private authenticationService: AuthenticationService,
