@@ -1,5 +1,5 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import {MatDialog} from "@angular/material";
+import { MatDialog } from "@angular/material/dialog";
 import {Router} from "@angular/router";
 import { LpNavMenuComponent } from './lp-nav-menu.component';
 import {MaterialModule} from "../../../material.module";
