@@ -1,7 +1,7 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
 import {AuthenticationService} from "../../../core/services/authentication.service";
 import {NavigationEnd, Router} from "@angular/router";
-import {MatDialog} from "@angular/material";
+import { MatDialog } from "@angular/material/dialog";
 import {LpConfirmationDialogComponent} from "../../lp-confirmation-dialog/lp-confirmation-dialog.component";
 import {Subscription} from "rxjs";
 import {filter} from "rxjs/operators";
