@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {Observable} from  'rxjs';
-import {MatSnackBar} from '@angular/material';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import {environment} from '../../../environments/environment';
 import {HttpClient} from "@angular/common/http";
 import {HttpService} from "./http.service";
